+++ src/Ydb.php
@@ -107,10 +107,14 @@ class Ydb
         $this->database = $config['database'] ?? null;
         $this->iam_config = $config['iam_config'] ?? [];
 
-        if ($logger){
-            $this->logger = $logger;
+        if (!is_null($logger) && isset($config['logger'])){
+            throw new \Exception('Logger set in 2 places');
+        } else if (isset($config['logger'])) {
+            $this->setLogger($config['logger']);
+        } else if ($logger) {
+            $this->setLogger($logger);
         } else {
-            $this->logger = new NullLogger();
+            $this->setLogger(new NullLogger());
         }
 
         $this->retry = new Retry($this->logger);
@@ -351,4 +355,21 @@ class Ydb
     {
         return $this->discoveryInterval;
     }
+
+    /**
+     * @return LoggerInterface|NullLogger
+     */
+    public function getLogger()
+    {
+        return $this->logger;
+    }
+
+    /**
+     * @param LoggerInterface $logger
+     * @return void
+     */
+    protected function setLogger(LoggerInterface $logger){
+        $this->logger = $logger;
+    }
+
 }