+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * fix keepInCache param in YdbQuery
+* add logger as Ydb config
+* added snapshot mode in noninteractive transaction
 
 ## 1.12.0
 * added StaticAuthentication